# docs/swagger.yaml
basePath: /api/v1
definitions:
  Response:
    properties:
      code:
        description: 状态码
        example: 200
        type: integer
      data:
        description: 返回数据
      msg:
        description: 提示信息
        example: success
        type: string
    type: object
  model.Book:
    properties:
      author:
        example: 许式伟
        type: string
      createdAt:
        example: "2021-01-01 00:00:00"
        type: string
      id:
        example: 1
        format: int64
        type: integer
      price:
        example: 99.99
        type: number
      publishDate:
        example: "2019-01-01"
        type: string
      publisher:
        example: 电子工业出版社
        type: string
      stock:
        example: 100
        type: integer
      summary:
        example: Go语言编程是一本介绍Go语言的书籍，内容包括Go语言的基础知识、并发编程、网络编程、Web编程、数据库编程等。
        type: string
      title:
        example: Go语言编程
        type: string
      updatedAt:
        example: "2021-01-01 00:00:00"
        type: string
    required:
    - author
    - price
    - publishDate
    - publisher
    - stock
    - summary
    - title
    type: object
  model.Lending:
    properties:
      book:
        $ref: '#/definitions/model.Book'
      book_id:
        example: 1
        type: integer
      createdAt:
        example: "2021-01-01 00:00:00"
        type: string
      id:
        example: 1
        format: int64
        type: integer
      lend_time:
        type: string
      reader:
        $ref: '#/definitions/model.Reader'
      reader_id:
        example: 1
        type: integer
      return_time:
        example: "2021-01-01"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.LendingStatus'
        enum:
        - 1
        - 2
        - 3
        example: 1
      updatedAt:
        example: "2021-01-01 00:00:00"
        type: string
    required:
    - book_id
    - lend_time
    - reader_id
    - return_time
    - status
    type: object
  model.LendingStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - LendingStatusLending
    - LendingStatusReturned
    - LendingStatusViolation
  model.Reader:
    properties:
      createdAt:
        example: "2021-01-01 00:00:00"
        type: string
      gender:
        example: 男
        type: string
      id:
        example: 1
        format: int64
        type: integer
      key:
        example: "123456"
        type: string
      name:
        example: 张三
        type: string
      phone:
        example: "18888888888"
        type: string
      studentNo:
        example: 2018000000
        type: integer
      updatedAt:
        example: "2021-01-01 00:00:00"
        type: string
    required:
    - gender
    - name
    - phone
    type: object
...
  v1.LoginResp:
    properties:
      token:
        example: xxx
        type: string
    type: object
info:
  contact:
    name: wear工程师
  description: 图书管理系统API文档
  title: 图书管理系统API文档
  version: v1
paths:
  /admin/changePassword:
    post:
      consumes:
      - application/json
      description: 管理员修改密码
      parameters:
      - description: 旧密码
        in: query
        name: oldPassword
        required: true
        type: string
      - description: 新密码
        in: query
        name: newPassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 管理员修改密码
      tags:
      - 管理员
  /admin/login:
    post:
      consumes:
      - application/json
      description: 管理员登录
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.LoginResp'
              type: object
      summary: 管理员登录
      tags:
      - 管理员
  /book:
    delete:
      consumes:
      - application/json
      description: 删除书籍
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 删除书籍
      tags:
      - 书籍
    post:
      consumes:
      - application/json
      description: 新书入库 | 更新书籍信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/model.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 新书入库 | 更新书籍信息
      tags:
      - 书籍
  /book/get:
    get:
      consumes:
      - application/json
      description: 查询书籍详情
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Book'
              type: object
      summary: 查询书籍详情
      tags:
      - 书籍
  /book/list:
    get:
      consumes:
      - application/json
      description: 查询书籍列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Book'
                  type: array
              type: object
      summary: 查询书籍列表
      tags:
      - 书籍
  /lend/book:
    post:
      consumes:
      - application/json
      description: 借书
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书籍ID
        in: body
        name: bookId
        required: true
        schema:
          type: integer
      - description: 借阅时长（天）
        in: body
        name: days
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 借书
      tags:
      - 借阅
  /lending/detail:
    get:
      consumes:
      - application/json
      description: 借阅记录详情
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 借阅记录ID
        in: query
        name: lendingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Lending'
              type: object
      summary: 借阅记录详情
      tags:
      - 借阅
  /lending/list:
    get:
      consumes:
      - application/json
      description: 借阅记录
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 读者ID
        in: query
        name: readerId
        type: integer
      - description: 书籍ID
        in: query
        name: bookId
        type: integer
      - description: 学号
        in: query
        name: studentNo
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 读者姓名
        in: query
        name: readerName
        type: string
      - description: 书籍名称
        in: query
        name: bookName
        type: string
      - description: 借阅时间开始
        in: query
        name: from
        type: string
      - description: 借阅时间结束
        in: query
        name: to
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Lending'
                  type: array
              type: object
      summary: 借阅记录
      tags:
      - 借阅
  /lending/listByReader:
    get:
      consumes:
      - application/json
      description: 读者查询借阅记录
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书籍ID
        in: body
        name: bookId
        required: true
        schema:
          type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Lending'
                  type: array
              type: object
      summary: 读者查询借阅记录
      tags:
      - 借阅
...
  /return/book:
    post:
      consumes:
      - application/json
      description: 还书
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书籍ID
        in: body
        name: bookId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
      summary: 还书
      tags:
      - 借阅
swagger: "2.0"